+++ src/main/java/com/daitda/hubservice/domain/hubinventory/application/dto/command/DecreaseHubInventoryCommand.java
@@ -1,4 +1,4 @@
-package com.daitda.hubservice.domain.hubinventory.application.dto;
+package com.daitda.hubservice.domain.hubinventory.application.dto.command;
 
 import lombok.Builder;
 import lombok.Getter;
