+++ src/main/java/com/daitda/hubservice/domain/hubinventory/application/service/HubInventoryService.java
@@ -1,9 +1,9 @@
 package com.daitda.hubservice.domain.hubinventory.application.service;
 
-import com.daitda.hubservice.domain.hubinventory.application.dto.CreateHubInventoryCommand;
-import com.daitda.hubservice.domain.hubinventory.application.dto.DecreaseHubInventoryCommand;
-import com.daitda.hubservice.domain.hubinventory.application.dto.RestoreHubInventoryCommand;
-import com.daitda.hubservice.domain.hubinventory.application.dto.UpdateHubInventoryCommand;
+import com.daitda.hubservice.domain.hubinventory.application.dto.command.CreateHubInventoryCommand;
+import com.daitda.hubservice.domain.hubinventory.application.dto.command.DecreaseHubInventoryCommand;
+import com.daitda.hubservice.domain.hubinventory.application.dto.command.RestoreHubInventoryCommand;
+import com.daitda.hubservice.domain.hubinventory.application.dto.command.UpdateHubInventoryCommand;
 import com.daitda.hubservice.domain.hubinventory.presentation.dto.response.FindHubInventoryResponse;
 import com.daitda.hubservice.domain.hubinventory.presentation.dto.response.ListHubInventoryResponse;
 import com.daitda.hubservice.domain.hubinventory.domain.entity.HubInventory;
