+++ src/test/java/com/daitda/hubservice/hubinventory/service/HubInventoryServiceTest.java
@@ -1,6 +1,6 @@
 package com.daitda.hubservice.hubinventory.service;
 
-import com.daitda.hubservice.domain.hubinventory.application.dto.CreateHubInventoryCommand;
+import com.daitda.hubservice.domain.hubinventory.application.dto.command.CreateHubInventoryCommand;
 import com.daitda.hubservice.domain.hubinventory.application.service.HubInventoryService;
 import com.daitda.hubservice.domain.hubinventory.presentation.dto.response.FindHubInventoryResponse;
 import com.daitda.hubservice.domain.hubinventory.domain.entity.HubInventory;
